+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/model/entity/CartEntity.java
@@ -17,9 +17,9 @@ public class CartEntity {
 
     private String name;
 
-    private BigDecimal price;
+    private long price;
 
-    private BigDecimal total;
+    private long total;
 
     private Integer quantity;
