+++ src/main/java/com/example/udpm14sellcomputerpartsbackend/repository/CartRepository.java
@@ -14,7 +14,8 @@ public interface CartRepository extends JpaRepository<CartEntity, Long> {
     CartEntity findAllByUserIdAndProductId(Long id, Long productId);
 
     @Query("SELECT SUM(c.total) FROM CartEntity c WHERE c.userId=?1")
-    BigDecimal sumPrice(Long id);
+    long sumPrice(Long id);
+
 
     @Query("SELECT SUM(c.quantity) FROM CartEntity c WHERE c.userId=?1")
     Integer sumQuantity(Long id);
